"Mine" and "Stub"type category hack change
As we now extract all recipes from the database again, we need sharper tools for cutting out what we do and don't show on the site.

The "_mine" category is immediately pulled out of the categories and turned into a binary toggle (0 or 1, false or true).

We also pull out the "_stub" and set the recipe type with it.